third_party/blink/web_tests/wpt_internal/css/css-contain/parsing/contain-valid.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Containment Module: parsing contain with valid values</title>
<!-- Merge these tests into upstream wpt test if/when added to the css-contain spec -->
<link rel="help" href="https://drafts.csswg.org/css-contain/#contain-property">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/1031">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/5796">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_valid_value("contain", "inline-size");
test_valid_value("contain", "block-size");
test_valid_value("contain", "layout inline-size", "inline-size layout");
</script>
